+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ## Templates
 ### Incidental Occurrence v3.0.0
 #### [Schema v3.0.0](https://github.com/gaiaresources/abis-mapping/blob/main/abis_mapping/templates/incidental_occurrence_data_v3/schema.json) changes (by column order).
+* The field `eventDate` is renamed `eventDateStart`.
+* The field `eventDateEnd` is added. It is an optional field with type Timestamp.
 * The field `conservationJurisdiction` is replaced by `conservationAuthority`.
 * `sensitivityCategory` and `sensitivityAuthority` are added fields. Both are type string.
 They are mutually inclusive, both must be provided or both must be blank.